Refactor transform
